pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "padelwatcher"
version = "1.0.0"
description = "Comprehensive padel court availability tracking and notification system"
readme = "README.md"
requires-python = ">=3.12,<4.0"
keywords = ["flask", "api", "backend", "padel", "court", "availability", "tracking"]
license = {text = "MIT"}
authors = [
    {name = "Thom Ezechiels", email = "thom@techletes.ai"},
]
maintainers = [
  { name = "Thom Ezechiels", email = "thom@techletes.ai" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Environment :: Web Environment",
    "Framework :: Flask",
]

# Add project dependencies here. pip-tools will read them from here, do not use a requirements file!
dependencies = [
    # Core Flask
    "Flask>=3.1",
    "flask-cors>=6.0",

    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.17",
    "psycopg2-binary>=2.9",

    # Authentication & Security
    "PyJWT>=2.10",
    "werkzeug==3.1",

    # API & Data Processing
    "httpx==0.28",
    "beautifulsoup4>=4.14",
    "selectolax>=0.4",
    "pydantic>=2.12",
    "gunicorn>=23.0",

    # Scheduling
    "apscheduler>=3.11",

    # Utilities
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-flask>=1.3.0",

    # Code quality
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",

    # Development
    "python-dotenv>=1.0.0",
    "ipython>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/thom-techlete/padelwatcher"
Repository = "https://github.com/thom-techlete/padelwatcher"
Issues = "https://github.com/thom-techlete/padelwatcher/issues"
Documentation = "https://github.com/thom-techlete/padelwatcher/blob/main/README.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "backend*"]
exclude = ["backend/tests*", "web*", "scripts*"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "--strict-markers",
    "--strict-config",
]
markers = [
    "coverage: mark test to run with coverage collection",
]
testpaths = ["**/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]

[tool.coverage.run]
source = ["app", "backend"]
omit = [
    "*/__pycache__/*",
    "*/tests/*",
    "*/.venv/*",
    "*/venv/*",
    "*/__init__.py",
    "setup.py",
    "*/alembic/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "except ImportError",
    "pass",
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
follow_imports = "silent"
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
no_implicit_optional = true
allow_redefinition = true
show_error_codes = true
show_column_numbers = true
pretty = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

exclude = [
  "alembic",
  "migrations",
  "tests",
  ".venv",
]

[tool.ruff]
target-version = "py312"
line-length = 88
fix = true
respect-gitignore = true
extend-exclude = ["alembic", "migrations", "*.ipynb"]

lint.select = [
  "E",     # pycodestyle errors
  "W",     # pycodestyle warnings
  "F",     # pyflakes
  "I",     # isort
  "B",     # flake8-bugbear
  "UP",    # pyupgrade
]

lint.ignore = [
  "E501",  # Line too long (handled by Black)
  "B008",  # Do not perform function calls in argument defaults
]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "alembic"]
skips = ["B101"]
